requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
pingouin>=0.5.4
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.4.0
openai>=1.30.0
python-dotenv>=1.0.0